% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_grm.R
\name{ccc_grm}
\alias{ccc_grm}
\title{Plot CCCs using the GRM}
\usage{
ccc_grm(theta_range = c(-5, 5), alpha, delta)
}
\arguments{
\item{theta_range}{Numeric vector of length 2, range of x-/theta-axis}

\item{alpha}{Numeric, item discrimination parameter}

\item{delta}{Numeric vector, threshold parameters}
}
\value{
A ggplot2 object
}
\description{
Plot CCCs using the GRM
}
\examples{
ccc_grm(alpha = 1.2, delta = c(-2, -0.5, 1.5))
}
