% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_2pl.R
\name{ci_2pl}
\alias{ci_2pl}
\title{Compute a confidence interval around a person parameter using the 2PL model}
\usage{
ci_2pl(theta, alpha, beta, ci_alpha, ci_direction)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{alpha}{Numeric, item discrimination parameter}

\item{beta}{Numeric, item difficulty}

\item{ci_alpha}{Number between 0 and 1, alpha-level}

\item{ci_direction}{Character, either "lower", "upper", or "both"}
}
\value{
Two-element numeric vector, limits of the confidence interval
}
\description{
If beta and alpha are vectors, the test information is used for the 
confidence interval.
}
\examples{
ci_2pl(
  theta = 0.5,
  alpha = c(0.5, 1.2, 1),
  beta = c(-0.5, 1, 2),
  ci_alpha = 0.05,
  ci_direction = "both"
)
}
