% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_3pl.R
\name{ci_3pl}
\alias{ci_3pl}
\title{Compute a confidence interval around a person parameter using the 3PL model}
\usage{
ci_3pl(theta, alpha, beta, gamma, ci_alpha, ci_direction)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{alpha}{Numeric, item discrimination parameter}

\item{beta}{Numeric, item difficulty}

\item{gamma}{Numeric, guessing parameter}

\item{ci_alpha}{Number between 0 and 1, alpha-level}

\item{ci_direction}{Character, either "lower", "upper", or "both"}
}
\value{
Two-element numeric vector, limits of the confidence interval
}
\description{
If beta, alpha, and gamma are vectors, the test information is used for the 
confidence interval.
}
\examples{
ci_3pl(
  theta = 0.5,
  alpha = c(1.2, 0.5),
  beta = c(-0.5, 1),
  gamma = c(0.3, 0.1),
  ci_alpha = 0.05,
  ci_direction = "both"
)
}
