% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_general.R
\name{ci_general}
\alias{ci_general}
\title{Compute confidence intervals around person parameters}
\usage{
ci_general(theta, inf, ci_alpha, ci_direction)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{inf}{Number, IRT information}

\item{ci_alpha}{Number between 0 and 1, alpha-level}

\item{ci_direction}{Character, either "lower", "upper", or "both"}
}
\value{
Numeric, two-element vector of confidence interval limits
}
\description{
This function is called by the specialized ci-functions.
}
