% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_rasch.R
\name{ci_rasch}
\alias{ci_rasch}
\title{Compute a confidence interval around a person parameter using the Rasch model}
\usage{
ci_rasch(theta, beta, ci_alpha, ci_direction)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{beta}{Numeric, item difficulty}

\item{ci_alpha}{Number between 0 and 1, alpha-level}

\item{ci_direction}{Character, either "lower", "upper", or "both"}
}
\value{
Two-element numeric vector, limits of the confidence interval
}
\description{
If beta is a vector, the test information is used for the confidence
interval.
}
\examples{
ci_rasch(
  theta = 0.5,
  beta = c(-1.2, 0, 2),
  ci_alpha = 0.05,
  ci_direction = "both"
)
}
