% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_z.R
\name{get_z}
\alias{get_z}
\title{Get z-values for a given alpha-level and test direction}
\usage{
get_z(ci_alpha, ci_direction)
}
\arguments{
\item{ci_alpha}{Number between 0 and 1, alpha-level}

\item{ci_direction}{Character, either "lower", "upper", or "both"}
}
\value{
Two-elements numeric vector, lower and upper z-value
}
\description{
ci_alpha is the alpha-level and should not be confused with the 
discrimination parameter used in other functions.
}
\examples{
get_z(ci_alpha = 0.05, ci_direction = "both")

}
