% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_2pl.R
\name{icc_2pl}
\alias{icc_2pl}
\title{Plot an ICC using the 2PL model}
\usage{
icc_2pl(theta_range = c(-5, 5), alpha, beta)
}
\arguments{
\item{theta_range}{Numeric vector of length 2, range of x-/theta-axis}

\item{alpha}{Numeric, item discrimination parameter}

\item{beta}{Numeric, item difficulty}
}
\value{
A ggplot2 object
}
\description{
Plot an ICC using the 2PL model
}
\examples{
icc_2pl(alpha = 1.2, beta = 1.5)
icc_2pl(alpha = c(1, 1.2), beta = c(0, 1.5))
}
