% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_3pl.R
\name{icc_3pl}
\alias{icc_3pl}
\title{Plot an ICC using the 3PL model}
\usage{
icc_3pl(theta_range = c(-5, 5), alpha, beta, gamma)
}
\arguments{
\item{theta_range}{Numeric vector of length 2, range of x-/theta-axis}

\item{alpha}{Numeric, item discrimination parameter}

\item{beta}{Numeric, item difficulty}

\item{gamma}{Numeric, guessing parameter}
}
\value{
A ggplot2 object
}
\description{
Plot an ICC using the 3PL model
}
\examples{
icc_3pl(alpha = 1, beta = 0.5, gamma = 0.3)
}
