% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_pcm.R
\name{icc_pcm}
\alias{icc_pcm}
\title{Plot ICCs using the PCM}
\usage{
icc_pcm(theta_range = c(-5, 5), delta)
}
\arguments{
\item{theta_range}{Numeric vector of length 2, range of x-/theta-axis}

\item{delta}{Numeric vector, threshold parameters}
}
\value{
A ggplot2 object
}
\description{
Plot ICCs using the PCM
}
\examples{
icc_pcm(delta = c(-1, 0, 1))
}
