% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_rasch.R
\name{icc_rasch}
\alias{icc_rasch}
\title{Plot an ICC using the Rasch model}
\usage{
icc_rasch(theta_range = c(-5, 5), beta)
}
\arguments{
\item{theta_range}{Numeric vector of length 2, range of x-/theta-axis}

\item{beta}{Numeric, item difficulty}
}
\value{
A ggplot2 object
}
\description{
Plot an ICC using the Rasch model
}
\examples{
icc_rasch(beta = -0.7)
icc_rasch(beta = 0:2)
}
