% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_tirt.R
\name{icc_tirt}
\alias{icc_tirt}
\title{Plot an ICC using the Thurstonian IRT model}
\usage{
icc_tirt(theta_range = c(-5, 5), mu, lambda, psi2, rotation = -210)
}
\arguments{
\item{theta_range}{Numeric vector of length 2, range of x-/theta-axis}

\item{mu}{Numeric vector of length 2, item means}

\item{lambda}{Numeric vector of length 2, factor loadings}

\item{psi2}{Numeric vector of length 2, residual variance (squared values)}

\item{rotation}{Numeric, increase or decrease to rotate}
}
\value{
Plots a 3D surface plot of the TIRT's ICC
}
\description{
Plot an ICC using the Thurstonian IRT model
}
\examples{
icc_tirt( 
  mu = c(0, -0.2), 
  lambda = c(0.7, 0.8), 
  psi2 = c(1, 1)
  )

}
