% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_2pl.R
\name{inf_2pl}
\alias{inf_2pl}
\title{Compute information at theta using the 2PL model}
\usage{
inf_2pl(theta, alpha, beta)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{alpha}{Numeric, item discrimination}

\item{beta}{Numeric, item difficulty}
}
\value{
Numeric, information
}
\description{
Compute information at theta using the 2PL model
}
\examples{
inf_2pl(theta = 0.5, alpha = 1.2, beta = 1.5)
}
