% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_3pl.R
\name{inf_3pl}
\alias{inf_3pl}
\title{Compute information at theta using the 3PL model}
\usage{
inf_3pl(theta, alpha, beta, gamma)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{alpha}{Numeric, item discrimination parameter}

\item{beta}{Numeric, item difficulty}

\item{gamma}{Numeric, guessing parameter}
}
\value{
Numeric, information
}
\description{
Compute information at theta using the 3PL model
}
\examples{
inf_3pl(theta = 0, alpha = 1, beta = 0.5, gamma = 0.3)
}
