% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_plot_2pl.R
\name{inf_plot_2pl}
\alias{inf_plot_2pl}
\title{Plot information function using the 2PL model}
\usage{
inf_plot_2pl(theta_range = c(-5, 5), alpha, beta, y_lim = NULL)
}
\arguments{
\item{theta_range}{Numeric vector of length 2, range of x-/theta-axis}

\item{alpha}{Numeric, item discrimination parameter}

\item{beta}{Numeric, item difficulty}

\item{y_lim}{Numeric, upper limit of the y-/information-axis}
}
\value{
A ggplot2 object
}
\description{
Plot information function using the 2PL model
}
\examples{
inf_plot_2pl(alpha = 1.2, beta = 1.5)
}
