% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_plot_rasch.R
\name{inf_plot_rasch}
\alias{inf_plot_rasch}
\title{Plot information function using the Rasch model}
\usage{
inf_plot_rasch(theta_range = c(-5, 5), beta)
}
\arguments{
\item{theta_range}{Numeric vector of length 2, range of x-/theta-axis}

\item{beta}{Numeric, item difficulty}
}
\value{
A ggplot2 object
}
\description{
Plot information function using the Rasch model
}
\examples{
inf_plot_rasch(beta = 0.5)
}
