% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_to_se.R
\name{inf_to_se}
\alias{inf_to_se}
\title{Convert IRT information to standard errors}
\usage{
inf_to_se(inf)
}
\arguments{
\item{inf}{Number, IRT information}
}
\value{
Numeric, standard error
}
\description{
Convert IRT information to standard errors
}
\examples{
inf_to_se(0.5)

}
