% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_2pl.R
\name{p_2pl}
\alias{p_2pl}
\title{Compute the probability to solve an item using the 2PL model}
\usage{
p_2pl(theta, alpha, beta)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{alpha}{Numeric, item discrimination}

\item{beta}{Numeric, item difficulty}
}
\value{
A numeric probability to solve the item
}
\description{
Compute the probability to solve an item using the 2PL model
}
\examples{
p_2pl(theta = 1.5, alpha = 1, beta = 0.5)

}
