% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_3pl.R
\name{p_3pl}
\alias{p_3pl}
\title{Compute the probability to solve an item using the 3PL model}
\usage{
p_3pl(theta, alpha, beta, gamma)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{alpha}{Numeric, item discrimination}

\item{beta}{Numeric, item difficulty}

\item{gamma}{Numeric, guessing parameter}
}
\value{
A numeric probability to solve the item
}
\description{
Compute the probability to solve an item using the 3PL model
}
\examples{
p_3pl(theta = -5, alpha = 1, beta = 0.5, gamma = 0.3)

}
