% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_gpcm.R
\name{p_gpcm}
\alias{p_gpcm}
\title{Compute category probabilities using the GPCM}
\usage{
p_gpcm(theta, alpha, delta)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{alpha}{Numeric, item discrimination parameter}

\item{delta}{Numeric vector, threshold parameters}
}
\value{
A numeric vector of probabilities to respond in categories
c = 0, ..., m
}
\description{
Compute category probabilities using the GPCM
}
\examples{
p_gpcm(theta = 0, alpha = 2, delta = c(-3, -1, 2))

}
