% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_grm.R
\name{p_grm}
\alias{p_grm}
\title{Compute category probabilities using the GRM}
\usage{
p_grm(theta, alpha, delta)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{alpha}{Numeric, item discrimination parameter}

\item{delta}{Numeric vector, threshold parameters}
}
\value{
A numeric vector of probabilities to respond in categories
c = 0, ..., m
}
\description{
Compute category probabilities using the GRM
}
\examples{
p_grm(theta = 1, alpha = 1.2, delta = c(-2, -0.5, 1.5))

}
