% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_pcm.R
\name{p_pcm}
\alias{p_pcm}
\title{Compute category probabilities using the PCM}
\usage{
p_pcm(theta, delta)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{delta}{Numeric vector, threshold parameters}
}
\value{
A numeric vector of probabilities to respond in categories
c = 0, ..., m
}
\description{
Compute category probabilities using the PCM
}
\examples{
p_pcm(theta = 0.5, delta = c(-1, 0, 1))

}
