% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_rasch.R
\name{p_rasch}
\alias{p_rasch}
\title{Compute the probability to solve an item using the Rasch model}
\usage{
p_rasch(theta, beta)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{beta}{Numeric, item difficulty}
}
\value{
A numeric probability to solve the item
}
\description{
Compute the probability to solve an item using the Rasch model
}
\examples{
p_rasch(theta = 3, beta = 2)
}
