% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_step_grm.R
\name{p_step_grm}
\alias{p_step_grm}
\title{Compute the probability respond in category c or above using a GRM}
\usage{
p_step_grm(theta, alpha, delta)
}
\arguments{
\item{theta}{Numeric, latent ability}

\item{alpha}{Numeric, item discrimination}

\item{delta}{Numeric vector, threshold parameters}
}
\value{
A numeric vector of probabilities to respond in category
c = 0, ..., m or above
}
\description{
Compute the probability respond in category c or above using a GRM
}
\examples{
p_step_grm(theta = 1, alpha = 1.2, delta = c(-2, -0.5, 1.5))

}
