% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_tirt.R
\name{p_tirt}
\alias{p_tirt}
\title{Compute the blockwise dependent probability to prefer the first item over the 
second in a forced-choice pair using the Thurstonian IRT (TIRT) model}
\usage{
p_tirt(theta, mu, lambda, psi2)
}
\arguments{
\item{theta}{Numeric vector of length 2, latent ability}

\item{mu}{Numeric vector of length 2, item means}

\item{lambda}{Numeric vector of length 2, factor loadings}

\item{psi2}{Numeric vector of length 2, residual variance (squared values)}
}
\value{
A numeric probability to prefer the first item over the second
}
\description{
Compute the blockwise dependent probability to prefer the first item over the 
second in a forced-choice pair using the Thurstonian IRT (TIRT) model
}
\examples{
p_tirt(
  theta = c(0, 0), 
  mu = c(0, -0.2), 
  lambda = c(0.7, 0.8), 
  psi2 = c(1, 1)
  )
}
