% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_2pl.R
\name{sim_2pl}
\alias{sim_2pl}
\title{Sample responses from a 2PL model}
\usage{
sim_2pl(n, m)
}
\arguments{
\item{n}{Numeric, number of participants}

\item{m}{Numeric, number of items}
}
\value{
A data.frame of simulated responses with \eqn{n} rows and \eqn{m} 
colums
}
\description{
Sample responses from a 2PL model
}
