% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_alpha.R
\name{sim_alpha}
\alias{sim_alpha}
\title{Sample m item discrimination parameters}
\usage{
sim_alpha(m)
}
\arguments{
\item{m}{Numeric, number of items}
}
\value{
A vector of item discrimination parameters
}
\description{
All items are assumed to be positively keyed.
}
\details{
\eqn{\boldsymbol{\alpha} \sim N(1, .2)}
}
