% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_beta.R
\name{sim_beta}
\alias{sim_beta}
\title{Sample m item difficulty parameters}
\usage{
sim_beta(m)
}
\arguments{
\item{m}{Numeric, number of items}
}
\value{
A vector of item difficulty parameters
}
\description{
\eqn{\boldsymbol{\beta} \sim N(0, 1)}
}
