% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_delta.R
\name{sim_delta}
\alias{sim_delta}
\title{Sample m_star threshold parameters}
\usage{
sim_delta(m_star)
}
\arguments{
\item{m_star}{Number of categories - 1}
}
\value{
A vector of sorted threshold parameters
}
\description{
Can be used for PCM, GPCM, and GRM
}
\details{
A vector of threshold differences is drawn. The distribution depends on the
number of categories to ensure an equal spacing of thresholds along the
standard normal metric of the person parameters. The threshold differences
are then centered around a mean item position that is sampled using
\eqn{\mu \sim N(0, .1)}.
}
