% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gamma.R
\name{sim_gamma}
\alias{sim_gamma}
\title{Sample m guessing parameters}
\usage{
sim_gamma(m)
}
\arguments{
\item{m}{Numeric, number of items}
}
\value{
A vector of guessing parameters
}
\description{
\eqn{\boldsymbol{\gamma} \sim \text{beta}(1, 5)}
}
