% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_grm.R
\name{sim_grm}
\alias{sim_grm}
\title{Sample responses from a Graded Response Model}
\usage{
sim_grm(n, m, m_star)
}
\arguments{
\item{n}{Numeric, number of participants}

\item{m}{Numeric, number of items}

\item{m_star}{Number of categories - 1}
}
\value{
A data.frame of simulated responses with \eqn{n} rows and \eqn{m} 
colums
}
\description{
Sample responses from a Graded Response Model
}
