% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_rasch.R
\name{sim_rasch}
\alias{sim_rasch}
\title{Sample responses from a Rasch model}
\usage{
sim_rasch(n, m)
}
\arguments{
\item{n}{Numeric, number of participants}

\item{m}{Numeric, number of items}
}
\value{
A data.frame of simulated responses with \eqn{n} rows and \eqn{m} 
colums
}
\description{
Sample responses from a Rasch model
}
