% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_theta.R
\name{sim_theta}
\alias{sim_theta}
\title{Sample n person parameters}
\usage{
sim_theta(n)
}
\arguments{
\item{n}{Numeric, number of participants}
}
\value{
A vector of person parameters
}
\description{
\eqn{\boldsymbol{\theta} \sim N(0, 1)}
}
