% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inf_plot_3pl.R
\name{inf_plot_3pl}
\alias{inf_plot_3pl}
\title{Plot information function using the 3PL model}
\usage{
inf_plot_3pl(theta_range = c(-5, 5), alpha, beta, gamma, y_lim = NULL)
}
\arguments{
\item{theta_range}{Numeric vector of length 2, range of x-/theta-axis}

\item{alpha}{Numeric, item discrimination}

\item{beta}{Numeric, item difficulty}

\item{gamma}{Numeric, guessing parameter}
}
\value{
A ggplot2 object
}
\description{
Plot information function using the 3PL model
}
\examples{
inf_plot_3pl(alpha = 1, beta = 0.5, gamma = 0.3)
}
